Refactor PostgreSQL tests to enhance SQL injection protection and add new test cases for DeleteRow, ListColumns, Query, and UpdateRow components
