+++ package.json
@@ -1,8 +1,8 @@
 {
     "scripts": {
         "lint": "eslint . --ext .js --no-error-on-unmatched-pattern",
-        "test-unit": "mocha --recursive --exit \"test/**/*.test.js\" \"src/appmixer/**/artifacts/test/**/*.test.js\"",
-        "test-unit-coverage": "nyc --reporter=lcov --reporter=text-summary mocha --recursive --exit \"test/**/*.test.js\" \"src/appmixer/**/artifacts/test/**/*.test.js\""
+        "test-unit": "mocha --recursive --exit --exclude '**/node_modules/**' 'test/**/*.test.js' 'src/appmixer/**/artifacts/test/**/*.test.js'",
+        "test-unit-coverage": "nyc --reporter=lcov --reporter=text-summary mocha --recursive --exit --exclude '**/node_modules/**' 'test/**/*.test.js' 'src/appmixer/**/artifacts/test/**/*.test.js'"
     },
     "devDependencies": {
         "autodetect-decoder-stream": "^2.0.2",
@@ -26,7 +26,6 @@
         "nyc": "^17.1.0",
         "onedrive-api": "1.0.9",
         "pdfjs-dist": "^4.10.38",
-        "pg": "8.13.1",
         "qs": "^6.10.1",
         "request-promise": "^4.2.2",
         "short-uuid": "^5.2.0",
